// src/app/models/venta.ts
export class venta {
    idventa: number;
    tipoventa: string;
    tipopago: string;
    estadoreserva: string;
    post: string;
    factura: string;
    recibo: string;
    idproducto: number;
    categoria: string;
    nombreproducto: string;
    serie: string;
    cantidad: number;
    existencia: number;
    preciounitario: number;
    preciototal: number;
    descuento: number;
    idsucursal: number;
    iduseradd: number;
    dateadd: string;
    idusermod: number;
    datemod: string;
}